refactor: css cleanup
