+++ components/Login.js
@@ -3,7 +3,7 @@ import Router from "next/router";
 import { useState, useEffect } from "react";
 import { setCookies, removeCookies } from "cookies-next";
 
-import styles from "../styles/Temp.module.css";
+import styles from "../styles/Form.module.css";
 
 function Login() {
   const [data, setData] = useState({
