+++ styles/Form.module.css
@@ -13,9 +13,7 @@
   box-shadow: 2px 4px 8px rgba(0, 0, 0, 0.1);
   max-width: 600px;
 }
-.title {
-  margin: 0 0 1.5rem;
-}
+
 .form label {
   display: block;
   text-transform: uppercase;
@@ -60,4 +58,4 @@
   border-radius: 4px;
   box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.1);
   cursor: pointer;
-}
+}
