+++ styles/globalStyles.css
@@ -1,7 +1,29 @@
+html,
+body {
+  padding: 0;
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
+    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+}
+
+* {
+  box-sizing: border-box;
+} 
+
 .container {
   padding: 0 2rem;
 }
 
+.center {
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+}
 .main {
   min-height: 100vh;
   padding: 4rem 0;
@@ -12,7 +34,8 @@
   align-items: center;
 }
 
-.footer {
+/* DELETE THIS WHEN NEW FOOTER CSS HAVE BEEN TESTED */
+/* .footer {
   display: flex;
   flex: 1;
   padding: 2rem 0;
@@ -26,7 +49,9 @@
   justify-content: center;
   align-items: center;
   flex-grow: 1;
-}
+} */
+
+
 
 .title a {
   color: #0070f3;
@@ -56,15 +81,6 @@
   font-size: 1.5rem;
 }
 
-.code {
-  background: #fafafa;
-  border-radius: 5px;
-  padding: 0.75rem;
-  font-size: 1.1rem;
-  font-family: Menlo, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono,
-    Bitstream Vera Sans Mono, Courier New, monospace;
-}
-
 .grid {
   display: flex;
   align-items: center;
@@ -73,28 +89,38 @@
   max-width: 800px;
 }
 
-.card {
-  margin: 1rem;
-  padding: 1.5rem;
+.card {  
+  display: flex;
+  margin: 0.5rem;
+  padding: 0.5rem;  
   text-align: left;
   color: inherit;
   text-decoration: none;
-  border: 1px solid #eaeaea;
+  list-style-type: none;
+  border: 1px solid #a09e9efa;
   border-radius: 10px;
   transition: color 0.15s ease, border-color 0.15s ease;
-  max-width: 300px;
+  width: 100vh;
+  
 }
-
+.button-container {
+  display: flex;
+  width: 100%;
+  justify-content: right;
+  
+}
 .card:hover,
 .card:focus,
 .card:active {
   color: #0070f3;
   border-color: #0070f3;
+  cursor: pointer;
 }
 
 .card h2 {
-  margin: 0 0 1rem 0;
-  font-size: 1.5rem;
+  margin: 0 0 0 0;
+  font-size: 1rem;
+  min-width: 50%;
 }
 
 .card p {
@@ -103,11 +129,7 @@
   line-height: 1.5;
 }
 
-.logo {
-  height: 1em;
-  margin-left: 0.5rem;
-}
-
+/* Mobile friendly layout */
 @media (max-width: 600px) {
   .grid {
     width: 100%;
